Change EventNumber from int to long

+++ src/EventStore.ClientAPI.Embedded/ClientMessageConverterExtensions.cs
@@ -57,9 +57,10 @@ namespace EventStore.ClientAPI.Embedded
 
         private static ClientMessage.EventRecord ToClientMessageEventRecord(this EventRecord eventRecord)
         {
+            // TODO Hayley
             if (eventRecord == null) return null;
             return new ClientMessage.EventRecord(
-                eventRecord.EventStreamId, eventRecord.EventNumber,
+                eventRecord.EventStreamId, (int)eventRecord.EventNumber,
                 eventRecord.EventId.ToByteArray(), eventRecord.EventType, eventRecord.IsJson ? 1 : 0,
                 eventRecord.IsJson ? 1 : 0,
                 eventRecord.Data,
