+++ src/EventStore.Core/Data/StreamMetadata.cs
@@ -12,17 +12,17 @@ namespace EventStore.Core.Data
     {
         public static readonly StreamMetadata Empty = new StreamMetadata();
 
-        public readonly int? MaxCount;
+        public readonly long? MaxCount;
         public readonly TimeSpan? MaxAge;
         
-        public readonly int? TruncateBefore;
+        public readonly long? TruncateBefore;
         public readonly bool? TempStream;
 
         public readonly TimeSpan? CacheControl;
         public readonly StreamAcl Acl;
 
-        public StreamMetadata(int? maxCount = null, TimeSpan? maxAge = null,
-                              int? truncateBefore = null, bool? tempStream = null,
+        public StreamMetadata(long? maxCount = null, TimeSpan? maxAge = null,
+                              long? truncateBefore = null, bool? tempStream = null,
                               TimeSpan? cacheControl = null, StreamAcl acl = null)
         {
             if (maxCount <= 0)
